refactor-basic-functionalities-for-admin-drivers

+++ src/main/resources/fleet-core-frontend/src/app/components-2/admin/pages/my-drivers/driver-card/driver-card.component.ts
@@ -2,6 +2,10 @@ import { Component, Input } from '@angular/core';
 import { Driver } from 'src/app/components-2/auth/dto/Driver';
 import { MyDriversService } from '../services/my-drivers-service.service';
 import { Router } from '@angular/router';
+import { formatNumber } from '@angular/common';
+import { MatDialog } from '@angular/material/dialog';
+import { DeleteDriverDialogComponent } from '../dialogs/delete-driver-dialog/delete-driver-dialog.component';
+import { MatSnackBar } from '@angular/material/snack-bar';
 
 @Component({
   selector: 'driver-card',
@@ -13,23 +17,34 @@ export class DriverCardComponent {
   @Input() driver: Driver;
 
   constructor(private driversService: MyDriversService,
-    private router: Router) { }
+              private router: Router,
+              private dialog: MatDialog) { }
 
   deleteDriver() {
-    this.driversService.deleteDriver(this.driver.id).subscribe(
-      (response) => {
-        alert(response);
-        window.location.reload();
-      },
-      (error) => {
-        alert(error);
-      }
-    );
 
+    const dialogRef = this.dialog.open(DeleteDriverDialogComponent);
+
+    dialogRef.afterClosed().subscribe(result => {
+      if(result === 'Delete'){
+        this.driversService.deleteDriver(this.driver.id).subscribe(
+          (response) => {
+            window.location.reload();
+          },
+          (error) => {
+            alert(error);
+          }
+        );
+      }});
   }
 
   viewDriverDetails() {
     this.router.navigate(["admin-dashboard/driver-details", this.driver.id]);
+    
+
+  }
+
+  formatMilenage(milenage: number): string {
+    return formatNumber(milenage, 'de', '1.0-0');
   }
 
 }
